refact: move ActivityStarter from :platform:ui to :platform:starter
