+++ core/mastodon/build.gradle.kts
@@ -58,6 +58,7 @@ dependencies {
   implementation(project(":ext:coroutines"))
   implementation(project(":platform:autos"))
   implementation(project(":platform:cache"))
+  implementation(project(":platform:starter"))
   implementation(project(":platform:ui"))
   implementation(libs.android.browser)
   implementation(libs.android.room.ktx)
