+++ feature/composer/build.gradle.kts
@@ -31,6 +31,7 @@ dependencies {
   implementation(project(":core"))
   implementation(project(":core:sample"))
   implementation(project(":platform:autos"))
+  implementation(project(":platform:starter"))
   implementation(project(":platform:ui"))
   implementation(libs.android.compose.material.icons)
