+++ feature/search/src/main/java/com/jeanbarrossilva/orca/feature/search/SearchActivity.kt
@@ -19,8 +19,8 @@ import android.content.Context
 import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
 import com.jeanbarrossilva.orca.platform.ui.core.composable.ComposableActivity
-import com.jeanbarrossilva.orca.platform.ui.core.on
 import com.jeanbarrossilva.orca.std.injector.Injector
+import com.jeanbarrossilva.platform.starter.on
 
 class SearchActivity internal constructor() : ComposableActivity() {
   private val module by lazy { Injector.from<SearchModule>() }
