+++ platform/starter/src/main/java/com/jeanbarrossilva/platform/starter/Context.extensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023 Orca
+ * Copyright © 2024 Orca
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -13,17 +13,17 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.platform.ui.core.activity
+package com.jeanbarrossilva.platform.starter
 
-import android.os.Bundle
-import androidx.fragment.app.FragmentActivity
+import android.content.Context
 
 /**
- * [FragmentActivity] that notifies its listeners from the [ActivityStarter] of its initialization.
+ * Creates an [ActivityStarter] for the [StartableActivity], from which it can be set up and
+ * started.
+ *
+ * @param T [StartableActivity] whose start-up may be configured.
+ * @see ActivityStarter.start
  */
-abstract class StartableActivity : FragmentActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    ActivityStarter.notifyListenersOf(this)
-  }
+inline fun <reified T : StartableActivity> Context.on(): ActivityStarter<T> {
+  return ActivityStarter(this, T::class)
 }
