+++ platform/ui/src/main/java/com/jeanbarrossilva/orca/platform/ui/core/Context.extensions.kt
@@ -18,20 +18,7 @@ package com.jeanbarrossilva.orca.platform.ui.core
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import com.jeanbarrossilva.orca.platform.ui.core.activity.ActivityStarter
-import com.jeanbarrossilva.orca.platform.ui.core.activity.StartableActivity
 import java.net.URL
-
-/**
- * Creates an [ActivityStarter] for the [StartableActivity], from which it can be set up and
- * started.
- *
- * @param T [StartableActivity] whose start-up may be configured.
- * @see ActivityStarter.start
- */
-inline fun <reified T : StartableActivity> Context.on(): ActivityStarter<T> {
-  return ActivityStarter(this, T::class)
-}
 
 /**
  * Browses to the [url].
