+++ platform/ui/src/main/java/com/jeanbarrossilva/orca/platform/ui/core/lifecycle/CompleteLifecycleActivity.kt
@@ -17,10 +17,10 @@ package com.jeanbarrossilva.orca.platform.ui.core.lifecycle
 
 import android.os.Bundle
 import androidx.lifecycle.Lifecycle
-import com.jeanbarrossilva.orca.platform.ui.core.activity.StartableActivity
 import com.jeanbarrossilva.orca.platform.ui.core.lifecycle.state.CompleteLifecycleState
 import com.jeanbarrossilva.orca.platform.ui.core.lifecycle.state.compareTo
 import com.jeanbarrossilva.orca.platform.ui.core.lifecycle.state.next
+import com.jeanbarrossilva.platform.starter.StartableActivity
 
 /**
  * [StartableActivity] that provides a complete version of [Lifecycle.State], containing missing
